event: submit_event() should take a 32-bit value
This is just a bug, events have a 32-bit value associated with them, but
submit_event() takes a 16-bit value by mistake. This didn't matter for
the keyboard codes, as they are not that large, but for larger 32-bit
values they would get truncated.

+++ include/protura/event/protocol.h
@@ -40,7 +40,7 @@ struct event_queue {
         .flags = (flgs), \
     }
 
-void event_queue_submit_event(struct event_queue *, uint16_t type, uint16_t code, uint16_t value);
+void event_queue_submit_event(struct event_queue *, uint16_t type, uint16_t code, uint32_t value);
 
 /* Wrap this function with your open call, and pass it the intended event queue */
 int event_queue_open(struct file *, struct event_queue *);
